+++ models/community.go
@@ -3,12 +3,12 @@ package models
 import "time"
 
 type Community struct {
-	ID   int    `json:"id" db:"community_id"`
+	ID   int64  `json:"id" db:"community_id"`
 	Name string `json:"name" db:"community_name"`
 }
 
 type CommunityDetail struct {
-	ID           int        `json:"id" db:"community_id"`
+	ID           int64      `json:"id" db:"community_id"`
 	Name         string     `json:"name" db:"community_name"`
 	Introduction string     `json:"introduction" db:"introduction"`
 	CreateTime   *time.Time `json:"create_time" db:"create_time"`
